Move access classes handling to separate header
